@@ -9,8 +9,8 @@
     "axios": "^0.18.0",
     "cors": "^2.8.5",
     "express": "^4.16.4",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
+    "react": "16.8.2",
+    "react-dom": "16.8.2",
     "react-redux": "^6.0.0",
     "react-router": "^4.3.1",
     "react-router-dom": "^4.3.1",
@@ -45,7 +45,7 @@
   },
   "jest": {
     "transform": {
-      "^.+\\.jsx?$": "babel-jest"      
+      "^.+\\.jsx?$": "babel-jest"
     },
     "verbose": true,
     "testURL": "http://localhost/"
