@@ -1,27 +1,15 @@
-import React, { Component } from 'react'
+import React, { useState } from 'react'
 import { CircularProgress } from '@material-ui/core';
 
-export default class AsyncImg extends Component {
-    state = {
-        loaded: false
+export default function AsyncImg({ src, className }) {
+    const [loaded, setLoaded] = useState(false);
+    function onLoad() {
+        setLoaded(true);
     }
-    constructor(){
-        super();
-        this.onLoad = this.onLoad.bind(this);
-    }
-    onLoad(){
-        this.setState({
-            loaded: true
-        })
-    }
-    render() {
-        const { src, className } = this.props;
-        const { loaded } = this.state;
-        return (
-            <div>
-                <CircularProgress className={className} style={{ padding: '70px', display: loaded ? 'none' : 'inherit'}}/>
-                <img className={className} src={src} alt="" style={{display: !loaded ? 'none' : 'inherit'}} onLoad={this.onLoad}/>
-            </div>            
-        )
-    }
+    return (
+        <div>
+            <CircularProgress className={className} style={{ padding: '70px', display: loaded ? 'none' : 'inherit'}}/>
+            <img className={className} src={src} alt="" style={{display: !loaded ? 'none' : 'inherit'}} onLoad={onLoad}/>
+        </div>            
+    )
 }
